+++ src/com/walmartlabs/lacinia/pedestal/interceptors.clj
@@ -1,0 +1,34 @@
+; Copyright (c) 2018-present Walmart, Inc.
+;
+; Licensed under the Apache License, Version 2.0 (the "License")
+; you may not use this file except in compliance with the License.
+; You may obtain a copy of the License at
+;
+;     http://www.apache.org/licenses/LICENSE-2.0
+;
+; Unless required by applicable law or agreed to in writing, software
+; distributed under the License is distributed on an "AS IS" BASIS,
+; WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+; See the License for the specific language governing permissions and
+; limitations under the License.
+
+(ns com.walmartlabs.lacinia.pedestal.interceptors
+  "Common interceptors between the standard and subscriptions code paths."
+  {:added "0.13.0"}
+  (:require
+    [io.pedestal.interceptor :refer [interceptor]]))
+
+(defn inject-app-context-interceptor
+  "Adds a :lacinia-app-context key to the request, used when executing the query.
+
+  The provided app-context map is augmented with the request map, as key :request.
+
+  It is not uncommon to replace this interceptor with one that constructs
+  the application context dynamically; for example, to extract authentication information
+  from the request and expose that as app-context keys."
+  [app-context]
+  (interceptor
+    {:name ::inject-app-context
+     :enter (fn [context]
+              (assoc-in context [:request :lacinia-app-context]
+                        (assoc app-context :request (:request context))))}))